refactor: refactor `ops` and `data` into a new `request` module that holds the domain logic
