+++ curlz/src/curlz/cli/mod.rs
@@ -1,7 +1,7 @@
-pub mod commands;
 pub mod execute;
 mod header_args;
+pub mod sub_commands;
 
-pub use commands::*;
 pub use execute::execute;
 pub use header_args::HeaderArgs;
+pub use sub_commands::*;
