--- curlz/src/curlz/data/mod.rs
@@ -1,5 +1,0 @@
-mod bookmark;
-mod request;
-
-pub use bookmark::*;
-pub use request::*;
