+++ curlz/src/curlz/lib.rs
@@ -1,11 +1,13 @@
 pub mod cli;
-pub mod data;
 pub mod interactive;
 pub mod ops;
 pub mod template;
 pub mod utils;
 pub mod variables;
 pub mod workspace;
+
+// the main domain logic
+pub mod request;
 
 mod http_file;
 #[cfg(test)]
