+++ curlz/src/curlz/request/mod.rs
@@ -1,0 +1,2 @@
+pub mod bookmark;
+pub mod http;
